backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyPDF2


# tem esse tbm 
# pip install python-multipart PyPDF2